+++ src/lang/interpreter.cpp
@@ -13,7 +13,6 @@
 #include "lang/program_util.hpp"
 #include "lang/semantics.hpp"
 #include "oeis/oeis_sequence.hpp"
-#include "sys/file.hpp"
 #include "sys/log.hpp"
 #include "sys/setup.hpp"