+++ CHANGELOG.md
@@ -4,7 +4,7 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## Featues
 
-* Support for program directives
+* Support for directives and `prg` instruction
 
 ### Enhancements